requirements.txt
fastapi
uvicorn
pillow
rembg
requests